# plans/phase4/openapi_ireios4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Real Estate AI Lead Agent",
    "description": "Advanced client-grade AI assistant backend for real estate tracking, dynamic questioning, and intelligent lead evaluation.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/graph/neighborhood": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Graph Neighborhood",
        "description": "IREIOS 4.0 ego neighborhood for Sales Copilot embed.\n\nSoft-fails (HTTP 200, available=false) when Neo4j is down or FEATURE_GRAPH_VIZ=false.",
        "operationId": "graph_neighborhood_api_v1_graph_neighborhood_get",
        "parameters": [
          {
            "name": "lead_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Center lead id",
              "title": "Lead Id"
            },
            "description": "Center lead id"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/leads/{lead_id}/context": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Graph Lead Context",
        "description": "Graph context for a lead, tenant-scoped (404 if not owned).",
        "operationId": "graph_lead_context_api_v1_graph_leads__lead_id__context_get",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      },
      "APIKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "api_key"
      }
    }
  }
}